pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screenmind-todo"
version = "0.1.0"
description = "Local AI-assisted todo list that watches screen activity and infers tasks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "pydantic-settings>=2.2.1,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pillow>=10.3.0,<11.0.0",
  "mss>=9.0.1,<10.0.0",
  "pytesseract>=0.3.10,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "psutil>=5.9.8,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "ruff>=0.5.0,<1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
